Product Matching/requirements.txt
datasketch>=1.6.0
pymongo>=4.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=2.0.0
numpy>=1.24.0
rapidfuzz>=3.0.0
